refactor(frontend): raster colour map config
Configure raster colour maps as `config/hazards/color-maps`, for consistency with other layer configs.
