+++ frontend/src/config/view-layers/hazards/HazardHoverDescription.tsx
@@ -3,7 +3,8 @@ import { FC } from 'react';
 import { RasterHoverDescription } from 'lib/data-map/types';
 import { RasterHoverDescription as RasterTooltip } from 'map/tooltip/content/RasterHoverDescription';
 
-import { HAZARDS_METADATA, HAZARD_COLOR_MAPS } from './metadata';
+import * as HAZARD_COLOR_MAPS from './color-maps';
+import { HAZARDS_METADATA } from './metadata';
 
 export const HazardHoverDescription: FC<RasterHoverDescription> = ({ target, viewLayer }) => {
   const { label, dataUnit } = HAZARDS_METADATA[viewLayer.id];
