+++ frontend/src/config/view-layers/hazards/HazardLegend.tsx
@@ -3,7 +3,8 @@ import { FC } from 'react';
 import { RasterLegend } from 'map/legend/RasterLegend';
 import { ViewLayer } from 'lib/data-map/view-layers';
 
-import { HAZARD_COLOR_MAPS, HAZARDS_METADATA } from './metadata';
+import * as HAZARD_COLOR_MAPS from './color-maps';
+import { HAZARDS_METADATA } from './metadata';
 
 export const HazardLegend: FC<{ viewLayer: ViewLayer }> = ({ viewLayer }) => {
   const { id } = viewLayer;
