+++ frontend/src/config/view-layers/hazards/color-maps.ts
@@ -1,0 +1,19 @@
+export const fluvial = {
+  scheme: 'blues',
+  range: [0, 10],
+};
+
+export const coastal = {
+  scheme: 'greens',
+  range: [0, 10],
+};
+
+export const surface = {
+  scheme: 'purples',
+  range: [0, 10],
+};
+
+export const cyclone = {
+  scheme: 'reds',
+  range: [0, 75],
+};
