+++ frontend/src/config/view-layers/hazards/metadata.ts
@@ -17,24 +17,5 @@ export const HAZARDS_METADATA = {
   },
 };
 
-export const HAZARD_COLOR_MAPS = {
-  fluvial: {
-    scheme: 'blues',
-    range: [0, 10],
-  },
-  coastal: {
-    scheme: 'greens',
-    range: [0, 10],
-  },
-  surface: {
-    scheme: 'purples',
-    range: [0, 10],
-  },
-  cyclone: {
-    scheme: 'reds',
-    range: [0, 75],
-  },
-};
-
 export const HAZARDS_MAP_ORDER = ['cyclone', 'fluvial', 'surface', 'coastal'];
 export const HAZARDS_UI_ORDER = ['fluvial', 'surface', 'coastal', 'cyclone'];
